Refactor rewards script removing pre-rewards utils
